requirements.txt
flask
psycopg2-binary
requests
pytest